requirements.txt
flask==3.0.3
flask-cors==5.0.0
pymupdf==1.24.9  # Versi lebih baru dengan binary Windows
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
torch==2.3.0
supabase==2.7.0
python-dotenv==1.0.1
gunicorn==22.0.0
scikit-learn==1.5.0
joblib==1.4.2
python-multipart==0.0.9
requests==2.32.3